Changed borders behaviour
Previously, when the player collided with a border its movement direction (x or y) was inverted. If the player swiped fastly towards any border, them used to break trough that border. Now, I clamped the position to the borders so when you collide with them the player just stops moving